Split Android progress repository packages
