+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/progress/ProgressHistoryConstants.kt
@@ -1,0 +1,3 @@
+package com.flashcardsopensourceapp.data.local.repository.progress
+
+const val progressHistoryDayCount: Long = 140L
