+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/progress/inputs/ProgressClockSnapshot.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.repository.progress
+package com.flashcardsopensourceapp.data.local.repository.progress.inputs
 
 import com.flashcardsopensourceapp.data.local.repository.TimeProvider
 import java.time.Instant
