+++ apps/android/data/local/src/test/java/com/flashcardsopensourceapp/data/local/repository/progress/snapshots/ProgressReviewScheduleSnapshotTest.kt
@@ -1,14 +1,24 @@
-package com.flashcardsopensourceapp.data.local.repository.progress
+package com.flashcardsopensourceapp.data.local.repository.progress.snapshots
 
 import com.flashcardsopensourceapp.data.local.model.cloud.CloudAccountState
 import com.flashcardsopensourceapp.data.local.model.progress.ProgressReviewScheduleBucketKey
 import com.flashcardsopensourceapp.data.local.model.progress.ProgressSnapshotSource
+import com.flashcardsopensourceapp.data.local.repository.progress.createCloudSettings
+import com.flashcardsopensourceapp.data.local.repository.progress.createPendingCardUpsertOutboxEntry
+import com.flashcardsopensourceapp.data.local.repository.progress.createPendingReviewOutboxEntry
+import com.flashcardsopensourceapp.data.local.repository.progress.createPendingScheduleCreateCardUpsertOutboxEntry
+import com.flashcardsopensourceapp.data.local.repository.progress.createPendingScheduleDeleteCardUpsertOutboxEntry
+import com.flashcardsopensourceapp.data.local.repository.progress.createPendingScheduleReviewCardUpsertOutboxEntry
+import com.flashcardsopensourceapp.data.local.repository.progress.createReviewSchedule
+import com.flashcardsopensourceapp.data.local.repository.progress.createReviewScheduleCardDue
+import com.flashcardsopensourceapp.data.local.repository.progress.inputs.hasPendingProgressReviewScheduleCardChanges
+import com.flashcardsopensourceapp.data.local.repository.progress.startOfLocalDateMillisForTest
+import java.time.LocalDate
+import java.time.ZoneId
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
-import java.time.LocalDate
-import java.time.ZoneId
 
 class ProgressReviewScheduleSnapshotTest {
     @Test
